Update Useful+features+identification.py
Added a simple model (linear regression) to see how good days in the past can predict the mood of the future, by looking at the explained variance. The explained variance shown is probably an overstatement due to the correlations of the predictors over the days. Nonetheless it can be used to determine the number of days in the past we want to look at ;)