Add files via upload
My current best lambdamart model. Trained on 10.000 srch ids from the 'train_test' dataset :). max_depth = 6 and the rest of the hyperparameters are default. It obtained a NDCG score of .506 on the test set containing also 10.000 srch ids. Trained 100 estimators with a total duration of 45 min. 

Note. You can load the model with pickle